subviews order unit tests
